refactor: move Window unit to src/Window subdirectory
