+++ tests/CMakeLists.txt
@@ -21,7 +21,7 @@ target_sources(tests
         "${PROJECT_SOURCE_DIR}/includes/WriteBuffer.hpp"
         "${PROJECT_SOURCE_DIR}/src/Cursor/Cursor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Offset/Offset.hpp"
-        "${PROJECT_SOURCE_DIR}/includes/Window.hpp"
+        "${PROJECT_SOURCE_DIR}/src/Window/Window.hpp"
         "${PROJECT_SOURCE_DIR}/src/TerminalState/TerminalState.hpp"
         "${PROJECT_SOURCE_DIR}/includes/Editor.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/Terminal.hpp"
